Added the bomb
Added the bomb as a gameplay element.
The bomb is always in the center of the game.
It has some amount of health, which will decrease if it is touched by an enemy (not by a bullet, by an enemy!)
if the bomb has no health left, it will explode and the game is over. (featuring a little animation)

Also added a score, which is increased by killing enemies.

Also added the possibility to change the text color of a Button.

Adjusted TODO list accordingly